rename terminalTabs to useTerminalTabs
