+++ src/hooks/bottomPanelTabs/useTerminalTabs.ts
